move teams reducer to typescript
also fix `testMatch` for jest so it runs typescript tests
