+++ frontend/src/store/actionTypes.ts
@@ -96,11 +96,14 @@ export const ADD_TEAM = "ADD_TEAM"
 export const DELETE_TEAM = "DELETE_TEAM"
 export const SET_LOADING_TEAM = "SET_LOADING_TEAM"
 export const SET_LOADING_TEAM_MEMBERS = "SET_LOADING_TEAM_MEMBERS"
+export const SET_LOADING_TEAM_INVITES = "SET_LOADING_TEAM_INVITES"
 export const SET_TEAM_404 = "SET_TEAM_404"
 export const SET_TEAM_MEMBERS = "SET_TEAM_MEMBERS"
+export const SET_TEAM_INVITES = "SET_TEAM_INVITES"
 export const SET_TEAM_RECIPES = "SET_TEAM_RECIPES"
 export const SET_LOADING_TEAM_RECIPES = "SET_LOADING_TEAM_RECIPES"
 export const SET_UPDATING_USER_TEAM_LEVEL = "SET_UPDATING_USER_TEAM_LEVEL"
+export const SET_UPDATING_MEMBERSHIP = "SET_UPDATING_MEMBERSHIP"
 export const SET_USER_TEAM_LEVEL = "SET_USER_TEAM_LEVEL"
 export const SET_DELETING_MEMBERSHIP = "SET_DELETING_MEMBERSHIP"
 export const DELETE_MEMBERSHIP = "DELETE_MEMBERSHIP"
@@ -114,6 +117,7 @@ export const UPDATE_TEAM = "UPDATE_TEAM"
 export const SET_LOADING_TEAMS = "SET_LOADING_TEAMS"
 
 export const SET_COPYING_TEAM = "SET_COPYING_TEAM"
+export const SET_MOVING_TEAM = "SET_MOVING_TEAM"
 
 export const SET_INVITES = "SET_INVITES"
 export const SET_LOADING_INVITES = "SET_LOADING_INVITES"
