+++ frontend/src/store/reducers/socialAccounts.ts
@@ -2,19 +2,28 @@ import {
   SET_SOCIAL_ACCOUNT_CONNECTIONS,
   SET_SOCIAL_ACCOUNT_CONNECTION
 } from "../actionTypes"
+import { ISocialConnection } from "./user"
 
-const initialState = {
+export interface ISocialAccountsState {
+  readonly github: ISocialConnection | null
+  readonly gitlab: ISocialConnection | null
+}
+
+const initialState: ISocialAccountsState = {
   github: null,
   gitlab: null
 }
 
-export const socialAccounts = (state = initialState, action) => {
+export const socialAccounts = (
+  state: ISocialAccountsState = initialState,
+  action: any
+) => {
   switch (action.type) {
     case SET_SOCIAL_ACCOUNT_CONNECTIONS:
       return {
         ...state,
         ...action.val.reduce(
-          (a, { provider, id }) => ({ ...a, [provider]: id }),
+          (acc: any, { provider, id }: any) => ({ ...acc, [provider]: id }),
           {}
         )
       }
